Refactor dir structure
