+++ src/service.rs
@@ -11,6 +11,10 @@ use log::error;
 use anyhow::Result;
 use qjs_sys::{c, convert::DecodeFromJSValue, JsCode, Output};
 
+mod resource;
+
+pub(crate) use resource::{OwnedJsValue, Resource};
+
 pub(crate) type ServiceRef = Rc<Service>;
 pub(crate) type ServiceWeakRef = Weak<Service>;
 
@@ -66,10 +70,7 @@ impl Service {
 
     pub fn dup_value(&self, value: c::JSValue) -> OwnedJsValue {
         let value = unsafe { c::JS_DupValue(self.ctx, value) };
-        OwnedJsValue {
-            service: self.weak_self(),
-            value,
-        }
+        OwnedJsValue::new(value, self.weak_self())
     }
 
     pub fn free_value(&self, value: c::JSValue) {
@@ -130,52 +131,6 @@ impl Drop for Service {
             drop(Box::from_raw(pname));
             c::JS_FreeContext(self.ctx);
             c::JS_FreeRuntime(self.runtime);
-        }
-    }
-}
-
-pub struct OwnedJsValue {
-    service: ServiceWeakRef,
-    value: c::JSValue,
-}
-
-impl Drop for OwnedJsValue {
-    fn drop(&mut self) {
-        let Some(service) = self.service.upgrade() else {
-            return;
-        };
-        service.free_value(self.value);
-    }
-}
-
-impl Clone for OwnedJsValue {
-    fn clone(&self) -> Self {
-        self.dup()
-            .expect("Failed to dup JsValue, the service has been dropped")
-    }
-}
-
-impl OwnedJsValue {
-    pub fn dup(&self) -> Option<Self> {
-        let service = self.service.upgrade()?;
-        Some(service.dup_value(self.value))
-    }
-
-    pub fn value(&self) -> &c::JSValue {
-        &self.value
-    }
-}
-
-pub struct Resource {
-    pub js_value: OwnedJsValue,
-    cancel_token: Option<Box<dyn Any>>,
-}
-
-impl Resource {
-    pub fn new(js_value: OwnedJsValue, cancel_token: Option<Box<dyn Any>>) -> Self {
-        Self {
-            js_value,
-            cancel_token,
         }
     }
 }
